Address code review comments
Move ScopedDestroyLock class in its own dedicated source file.
