+++ framework/encode/vulkan_scoped_destroy_lock.cpp
@@ -1,0 +1,56 @@
+/*
+** Copyright (c) 2023 Advanced Micro Devices, Inc. All rights reserved.
+**
+** Permission is hereby granted, free of charge, to any person obtaining a
+** copy of this software and associated documentation files (the "Software"),
+** to deal in the Software without restriction, including without limitation
+** the rights to use, copy, modify, merge, publish, distribute, sublicense,
+** and/or sell copies of the Software, and to permit persons to whom the
+** Software is furnished to do so, subject to the following conditions:
+**
+** The above copyright notice and this permission notice shall be included in
+** all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+** IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+** FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+** AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+** LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+** FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+** DEALINGS IN THE SOFTWARE.
+*/
+
+#include "vulkan_scoped_destroy_lock.h"
+
+GFXRECON_BEGIN_NAMESPACE(gfxrecon)
+GFXRECON_BEGIN_NAMESPACE(encode)
+
+std::shared_mutex ScopedDestroyLock::mutex_for_create_destroy_handle_;
+
+ScopedDestroyLock::ScopedDestroyLock(bool shared)
+{
+    lock_shared_ = shared;
+    if (shared)
+    {
+        mutex_for_create_destroy_handle_.lock_shared();
+    }
+    else
+    {
+        mutex_for_create_destroy_handle_.lock();
+    }
+};
+
+ScopedDestroyLock::~ScopedDestroyLock()
+{
+    if (lock_shared_)
+    {
+        mutex_for_create_destroy_handle_.unlock_shared();
+    }
+    else
+    {
+        mutex_for_create_destroy_handle_.unlock();
+    }
+};
+
+GFXRECON_END_NAMESPACE(encode)
+GFXRECON_END_NAMESPACE(gfxrecon)
